+++ src/main/java/com/example/umc8th/domain/member/controller/MemberController.java
@@ -1,16 +1,13 @@
 package com.example.umc8th.domain.member.controller;
 
-import com.example.umc8th.domain.article.dto.response.MemberResponseDTO;
 import com.example.umc8th.domain.member.dto.request.MemberRequestDTO;
+import com.example.umc8th.domain.member.dto.response.MemberResponseDTO;
 import com.example.umc8th.domain.member.entity.Member;
 import com.example.umc8th.domain.member.service.command.MemberCommandService;
 import com.example.umc8th.global.apiPayload.CustomResponse;
 import com.example.umc8th.global.apiPayload.success.GeneralSuccessCode;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/auth")
@@ -23,5 +20,12 @@ public class MemberController {
     public CustomResponse<MemberResponseDTO.SignUpResponseDTO> signUp(@RequestBody MemberRequestDTO.SignUpRequestDTO dto) {
         Member member = memberCommandService.signUp(dto);
         return CustomResponse.onSuccess(GeneralSuccessCode.OK, MemberResponseDTO.SignUpResponseDTO.from(member));
+    }
+
+    @GetMapping("/login")
+    public CustomResponse<MemberResponseDTO.LoginResponseDTO> login(@RequestBody MemberRequestDTO.LoginRequestDTO dto) {
+        MemberResponseDTO.LoginResponseDTO member = memberCommandService.login(dto);
+
+        return CustomResponse.onSuccess(GeneralSuccessCode.OK, member);
     }
 }
