+++ src/main/java/com/example/umc8th/domain/member/dto/response/MemberResponseDTO.java
@@ -1,17 +1,33 @@
 package com.example.umc8th.domain.member.dto.response;
 
-import lombok.AllArgsConstructor;
+import com.example.umc8th.domain.member.entity.Member;
 import lombok.Builder;
-import lombok.Getter;
 
 public class MemberResponseDTO {
 
-    @Getter
     @Builder
-    @AllArgsConstructor
-    public static class LoginResponseDTO {
-        private Long id;
-        private String accessToken;
-        private String refreshToken;
+    public record SignUpResponseDTO (
+            Long id
+    ) {
+        public static SignUpResponseDTO from(Member member) {
+            return SignUpResponseDTO.builder()
+                    .id(member.getId())
+                    .build();
+        }
+    }
+
+    @Builder
+    public record LoginResponseDTO (
+            Long id,
+            String accessToken,
+            String refreshToken
+    ) {
+        public static LoginResponseDTO from(Member member) {
+            return LoginResponseDTO.builder()
+                    .id(member.getId())
+                    .accessToken(member.getAccessToken())
+                    .refreshToken(member.getRefreshToken())
+                    .build();
+        }
     }
 }
